ipn: make Notify.Prefs be a *ipn.PrefsView
It is currently a `ipn.PrefsView` which means when we do a JSON roundtrip,
we go from an invalid Prefs to a valid one.

This makes it a pointer, which fixes the JSON roundtrip.

This was introduced in 0957bc5af2a544f7d197ea1e743696e8ba103b9f.

Signed-off-by: Maisem Ali <maisem@tailscale.com>
(cherry picked from commit 6afe26575cf155bd0baba8e4770bc49c0d55e2c1)