ipn/ipnlocal: fix controlclient reentrancy when deleting profile during logout
Deletion of profiles on logout (#6297) added a LocalBackend.Start()
call within setClientStatus(), but that's a callback from the
controlclient. Start() ends calling back into the controlclient (to
shut it down), and we end up stuck in a deadlock waiting for the
authDone channel to be closed.

Fixed by making the Start call asynchronous. To reproduce this in a test
case, we need to do an asynchronous logout, so add a CLI option for
that.

Signed-off-by: Mihai Parparita <mihai@tailscale.com>