tsnet: fix handling of non-empty hostname in Server.Listen
If Listen is called with a non-empty hostname, packets will not correctly be
forwarded. Document that the hostname must be empty, and report an error if it
is not.

Since the hostname is now definitely empty, omit it from the listener's lookup
plumbing.

Moreover, since we only support TCP protocols, that does not need to be part of
the lookup key either.

Fixes #6815.

Change-Id: I01be0232f323f15d672bd78af52f7c6740c86c30
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>