allow tailscale us without iptables in system
nftables is here to stay and legacy iptables conversion tool does exist- but everyone does not want to use it.
Or their systems might not have it at all;
for example, openwrt routers have gone to nftables and deprecated iptables some time ago,
and some other systems also have gone that way- then there is problems with tailscaling...

note: it is possible to use iptables or iptables-nft in openwrt, just that as default, it has only nftables,
and some people might want to keep it that way.

There is a way, use --netfilter-mode=off to disable all netfilter related features,
except that it is not enough-
when you start server, tailscaled - it won't start if iptables
is not found from path - as it constructs ipt4 and init executes iptables
command with argument --version. Ofcourse this fails in a system without
iptables, as executable is not available.

Workaround: you could make a shell script to /usr/bin with name iptables,
and another possibly with ip6tables, if you are looking forward to have ipv6
support as well, then make it regocnize --version argument and echo a "fake"
iptables version string. But if your system runs other software that is capable
of using either, nftables or iptables, and prefers iptables, well- rules are
made by dummy iptables command. Instead we would need native nftables support
for tailscale, but most likely, it will take some more time until development
reaches that point. For now, we can use dummy iptables as netfilter, but it
has it's downsizes. This PR allows us to start tailscaled, without iptables.

Signed-off-by: Oskari Rauta <oskari.rauta@gmail.com>