disable netfilter if iptables is not present
on nftables or any other system using other firewall
than iptables, iptables command is not available
(unless nft-compat/iptables-nft is used)

On these systems, tailscaled cannot be used, as even with
tailscale up -netfilter-mode off, tailscaled daemon constructs
ipt4 and ipt6 - and if ipt4 fails, daemon won't start.

As a workaround, user can create /usr/bin/iptables and
/usr/bin/ip6tables shell scripts responding to --version
and if that argument is used, script returns faked
version string. Downside of this is that if system runs
other software capable of using iptables and prefers
to iptables, rules are made with faking shell script.

With this commit, netfilter mode is set off, in case
that iptables is not present in system and all functions
that handle rules are ignored, allowing users without
iptables to use tailscale. Also, if iptables is missing
SNAT rules are also ignored.

If user has iptables- it is used as this commit would
not exist at all, so it does not affect iptables users
in anyway.

Signed-off-by: Oskari Rauta <oskari.rauta@gmail.com>