.github/workflows: separate pre- and post-merge CI
The post-merge CI invokes the pre-merge CI workflow, then adds a
slack notification step.

This is because github won't collapse a finished CI into a single
"everything was fine" item if there are any skipped jobs. By having
a combined pre- and post- CI workflow with a skippable slack notification
step, the UX for PRs was quite a bit worse.

Signed-off-by: David Anderson <danderson@tailscale.com>