tool/gocross: always create a gocross overlay GOROOT
Some of our tests and build processes reinvoke go by using runtime.GOROOT()+"bin/go",
or need to get at files in GOROOT/src (e.g. for the wasm builds). gocross breaks
both of these because it always -trimpaths for reproducibility.

As a fix for that, have gocross create a GOROOT that is identical to the underlying
toolchain, except that bin/go is still gocross, and pass that GOROOT as an env var
to build-ish commands (test, run, etc.). This makes our build/test programs that
rely on GOROOT work as before, but still produces trimmed binaries that are
suitable for releasing.

Updates tailscale/corp#3531, tailscale/corp#3539

Signed-off-by: David Anderson <danderson@tailscale.com>